Show modal only on edit tab click and handle readonly versions
This commit changes the edit permissions handling, so that the
permissions modal is shown when the edit tab is clicked, instead of
being shown on navigating to /edit routes. This prevents the modal from
reappearing when pressing the back button (SWIK-1200). Also, if a user has
permissions to edit a deck but the revision selected is not the latest,
then the modal informs them accordingly.